added async execution, code enhancements, and region support
