+++ CHANGES
@@ -28,5 +28,11 @@
 1.5   01/27/2019 -- Added test_yf_dividend_price() unit testing method.
 1.6   10/18/2020 -- Merged in two pull requests with bug fixes from sylvandb.
 1.7   01/01/2023 -- Merged in pull request with fixes from sedwards2000.
-1.7   01/01/2023 -- Support for python 2.7 dropped.
-1.8   01/09/2023 -- Improved decryption support
+1.7   01/01/2023 -- Support for Python 2.7 dropped.
+1.8   01/09/2023 -- Improved decryption support.
+1.9   01/14/2023 -- Added official Python 3.11 support.
+1.9   01/14/2023 -- General code cleanup & quality improvements.
+1.9   01/14/2023 -- Added new optional concurrent input to YahooFinancials(), if True extraction will run async.
+1.9   01/14/2023 -- Added new optional meta input to YahooFinancials(), currently supports language & region.
+1.9   01/14/2023 -- Updated beautifulsoup4 find to use string parameter instead deprecated text parameter.
+
