+++ setup.py
@@ -10,11 +10,11 @@ with codecs.open('README.rst', encoding='utf-8') as f:
 
 setup(
     name='yahoofinancials',
-    version='1.8',
+    version='1.9',
     description='A powerful financial data module used for pulling both fundamental and technical data from Yahoo Finance',
     long_description=long_description,
     url='https://github.com/JECSand/yahoofinancials',
-    download_url='https://github.com/JECSand/yahoofinancials/archive/1.8.tar.gz',
+    download_url='https://github.com/JECSand/yahoofinancials/archive/1.9.tar.gz',
     author='Connor Sanders',
     author_email='connor@exceleri.com',
     license='MIT',
@@ -38,6 +38,7 @@ setup(
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10'
+        'Programming Language :: Python :: 3.11'
     ],
     zip_safe=False
 )
