+++ yahoofinancials/maps.py
@@ -1,0 +1,16 @@
+COUNTRY_MAP = {
+    "FR": {"lang": "fr-FR", "region": "FR", "corsDomain": "fr.finance.yahoo.com"},
+    "IN": {"lang": "en-IN", "region": "IN", "corsDomain": "in.finance.yahoo.com"},
+    "HK": {"lang": "zh-Hant-HK", "region": "HK", "corsDomain": "hk.finance.yahoo.com"},
+    "DE": {"lang": "de-DE", "region": "DE", "corsDomain": "de.finance.yahoo.com"},
+    "CA": {"lang": "en-CA", "region": "CA", "corsDomain": "ca.finance.yahoo.com"},
+    "ES": {"lang": "es-ES", "region": "ES", "corsDomain": "es.finance.yahoo.com"},
+    "IT": {"lang": "it-IT", "region": "IT", "corsDomain": "it.finance.yahoo.com"},
+    "US": {"lang": "en-US", "region": "US", "corsDomain": "finance.yahoo.com"},
+    "AU": {"lang": "en-AU", "region": "AU", "corsDomain": "au.finance.yahoo.com"},
+    "UK": {"lang": "en-GB", "region": "GB", "corsDomain": "uk.finance.yahoo.com"},
+    "BR": {"lang": "pt-BR", "region": "BR", "corsDomain": "br.financas.yahoo.com"},
+    "NZ": {"lang": "en-NZ", "region": "NZ", "corsDomain": "nz.finance.yahoo.com"},
+    "SG": {"lang": "en-SG", "region": "SG", "corsDomain": "sg.finance.yahoo.com"},
+    "TW": {"lang": "zh-tw", "region": "TW", "corsDomain": "tw.finance.yahoo.com"},
+}
